Update README.md
This program is used to check if the inputs given in a sudoku puzzle is correct and to identify the incorrect inputs.This code works till 9x9 matrix. Numbers of rows and columns are given in the input.Once the inputs are entered ,the numbers are displayed and checks for the inputs.If the inputs are correct it displays "correct'. If incorrect it specifies the position of incorrect input specifying the nth row and nth column.Thus it check for the correct inputs for a sudoku puzzle.